move list historical balance test under address
